backend_py/pyproject.toml
[project]
name = "backend-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "cloudinary>=1.44.1",
    "fastapi>=0.131.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "motor>=3.7.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "python-socketio>=5.16.1",
    "uvicorn[standard]>=0.41.0",
]
